backend/requirements.txt
asgiref 
django 
django-cors-headers
djangorestframework 
djangorestframework-simplejwt
django-dotenv
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
django-allauth
google-auth
google-auth-oauthlib
google-auth-httplib2
requests
gunicorn
